Footer Fix - fix mobile styles - move Footer to own dir
